job-search-app/backend/requirements.txt
# FastAPI core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Pydantic & settings
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic[email]==2.10.1

# AI
anthropic==0.40.0

# Export
markdown==3.7
weasyprint==63.0  # optional: PDF export
python-docx==1.1.2  # optional: DOCX export

# Utils
python-jose[cryptography]==3.3.0
httpx==0.27.2

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
